Pass the raquet tile dimensions to MEOS unnarrowed (#273)
`rasterTileValueQuadbin` took its width and height as INTEGER and narrowed
them to `uint16_t` at the call site.  MEOS takes those arguments in the wider
type precisely so that it can reject a value outside the tile's 16-bit range
before storing it, so narrowing first handed it an already-wrapped number that
its check could never see: a width of 65537 wrapped to 1 and the query
silently sampled a one-pixel-wide tile instead of failing.

    SELECT startValue(rasterTileValueQuadbin(from_hex('2a'), 65537, 1,
      quadbinTileToCell(0,0,0), 'UINT8', 0.0, false,
      tgeompoint 'POINT(0 0)@2024-01-01'));
    -- 42.0, the same answer a width of 1 gives

The dimensions now reach MEOS in the type the SQL surface uses, as the sibling
`raquet()` constructor in the same file already passed them, and the
out-of-range value is reported rather than wrapped.  The tests cover both
entry points at both ends of the range.

+++ src/raster/raquet.cpp
@@ -647,7 +647,7 @@ void RaquetFunctions::Raster_tile_value_quadbin(
         Temporal *res = raster_tile_value_quadbin(
             t, reinterpret_cast<const uint8_t *>(pixels[row].GetData()),
             static_cast<size_t>(pixels[row].GetSize()),
-            static_cast<uint16_t>(width[row]), static_cast<uint16_t>(height[row]),
+            width[row], height[row],
             static_cast<uint64_t>(cell[row]), pt,
             nodata[row], has_nodata[row]);
         free(t);
